replace 'go get' with 'go mod verify'
PR replaces no longer needed / redundant `go get ./...` invocation with a new `modverify` step running `go mod verify`.  It also simplifies the `gofmt` command in the corresponding step: it is useless to format the data (with `-w` flag) in the build since formatting is for humans, the compiler doesn't care much about whitespaces and so on 😄 On the other hand, I believe we should fail the build if the contributor hasn't done the proper formatting of the modified code.